requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask-wtf==1.1.1
flask-sqlalchemy==3.0.5
email-validator==2.0.0
pytest==7.4.0
flask-migrate==4.0.4
gunicorn==21.2.0
Werkzeug==2.3.7